Minor fix: Copy-paste issue in the list of used baselineLD annotations
The file mentions an annotation named "H3K4me1_Trynka.extend.50H3K4me3_peaks_Trynka". There is no such annotation in baselineLD, but there are two annotations called "H3K4me1_Trynka.extend.500" and "H3K4me3_peaks_Trynka" that go one after another in baselineLD list.  
I think that a glitch happened when the list of used annotations was copy-pasted to this repo after the analysis was done, causing a merger of the two sequential lines. It seems unlikely to me that this has affected the original analysis.  
This PR fixes this merger (and also changes the name of the file accordingly).